momentum-shift/backend/requirements.txt
# Phase 1: Environment Setup
fastapi
uvicorn
supabase
nba_api
pandas
numpy
scikit-learn
joblib
python-dotenv
httpx
